+++ Assets/Scripts/Collectibles/Coin.cs
@@ -6,7 +6,7 @@ namespace Collectibles {
 
         private void OnTriggerEnter2D(Collider2D other) {
             if (other.CompareTag("Player")) {
-                LevelController.instance.coinsCollected += 1;
+                GameController.instance.coinsCollected += 1;
                 HUDController.instance.CalcCoins();
                 Destroy(gameObject);
             }