+++ Assets/Scripts/Objects/Pot.cs
@@ -33,7 +33,7 @@ namespace Objects {
         }
 
         private void TakeDamage() {
-            health -= Controllers.LevelController.instance.playerDamage;
+            health -= Controllers.GameController.instance.playerDamage;
 
             if (health <= 0) {
                 anim.Play("Pot");