+++ Assets/Scripts/Utils/ResetRoomNum.cs
@@ -5,8 +5,8 @@ namespace Utils {
     public class ResetRoomNum : MonoBehaviour {
 
         public void ResetRoom() {
-            if (LevelController.instance.currentRoom > 8) {
-                LevelController.instance.currentRoom = 0;
+            if (GameController.instance.currentRoom > 8) {
+                GameController.instance.currentRoom = 0;
                 Debug.Log("resetting current room");
             }
         }